tensorflow/requirements.txt
image==1.5.5
matplotlib==2.0.2
numpy==1.13.2
opencv-python==3.3.0.10
Pillow==4.2.1
tensorflow==1.3.0
tensorflow-tensorboard==0.1.7